B-502a-fix (senior): temperature f32→f64 — undo the doctored round-trip fixture
B-502a's round-trip test passed only because the fixture hardcoded
temperature: 0.699999988079071 (the f32-rounded value of 0.7) to match the
lossy output — masking a real bug, not fixing it. Root cause: ADR-0005's
temperature: Option<f32>; JSON numbers are f64, so read→f32→write silently
mutates a caller's 0.7 → 0.699999988 on every same-protocol request once
write_request is wired in (B-503). That is exactly the LiteLLM-style lossiness
busbar exists to prevent.

Fix: IrRequest.temperature is Option<f64> (a deliberate correction to ADR-0005,
which was wrong to use f32 for a JSON-sourced number); read_request drops the
`as f32` cast; fixtures restored to the honest 0.7. Round-trip now passes for
the right reason. 107 tests, clippy clean.

+++ src/ir.rs
@@ -12,7 +12,9 @@ pub(crate) struct IrRequest {
     pub messages: Vec<IrMessage>,
     pub tools: Vec<IrTool>,
     pub max_tokens: Option<u32>,
-    pub temperature: Option<f32>,
+    // f64 (not ADR-0005's f32): JSON numbers are f64; an f32 round-trip silently mutates a
+    // caller's temperature (0.7 → 0.699999988) — the exact lossiness busbar exists to avoid.
+    pub temperature: Option<f64>,
     pub stream: bool,
     pub extra: serde_json::Map<String, serde_json::Value>,
 }
